AutomatedAITest/requirements.txt
grpcio>=1.62.0
grpcio-tools>=1.62.0
protobuf>=4.25.0
requests>=2.31.0
pandas>=2.2.0
